Migrate `update_boundary` to new trait
