+++ crates/fj-core/src/operations/geometry/half_edge.rs
@@ -1,6 +1,11 @@
+use fj_math::Point;
+
 use crate::{
-    geometry::SurfacePath, objects::HalfEdge, operations::insert::Insert,
-    storage::Handle, Core,
+    geometry::{CurveBoundary, SurfacePath},
+    objects::HalfEdge,
+    operations::insert::Insert,
+    storage::Handle,
+    Core,
 };
 
 /// Update the geometry of a [`HalfEdge`]
@@ -10,6 +15,14 @@ pub trait UpdateHalfEdgeGeometry {
     fn update_path(
         &self,
         update: impl FnOnce(SurfacePath) -> SurfacePath,
+        core: &mut Core,
+    ) -> Self;
+
+    /// Update the boundary of the edge
+    #[must_use]
+    fn update_boundary(
+        &self,
+        update: impl FnOnce(CurveBoundary<Point<1>>) -> CurveBoundary<Point<1>>,
         core: &mut Core,
     ) -> Self;
 }
@@ -25,6 +38,20 @@ impl UpdateHalfEdgeGeometry for Handle<HalfEdge> {
         HalfEdge::new(
             path,
             self.boundary(),
+            self.curve().clone(),
+            self.start_vertex().clone(),
+        )
+        .insert(core)
+    }
+
+    fn update_boundary(
+        &self,
+        update: impl FnOnce(CurveBoundary<Point<1>>) -> CurveBoundary<Point<1>>,
+        core: &mut Core,
+    ) -> Self {
+        HalfEdge::new(
+            self.path(),
+            update(self.boundary()),
             self.curve().clone(),
             self.start_vertex().clone(),
         )
