+++ crates/fj-core/src/operations/update/half_edge.rs
@@ -1,7 +1,4 @@
-use fj_math::Point;
-
 use crate::{
-    geometry::CurveBoundary,
     objects::{Curve, HalfEdge, Vertex},
     operations::{derive::DeriveFrom, insert::Insert},
     storage::Handle,
@@ -10,13 +7,6 @@ use crate::{
 
 /// Update a [`HalfEdge`]
 pub trait UpdateHalfEdge {
-    /// Update the boundary of the edge
-    #[must_use]
-    fn update_boundary(
-        &self,
-        update: impl FnOnce(CurveBoundary<Point<1>>) -> CurveBoundary<Point<1>>,
-    ) -> Self;
-
     /// Update the curve of the edge
     #[must_use]
     fn update_curve<T>(
@@ -39,18 +29,6 @@ pub trait UpdateHalfEdge {
 }
 
 impl UpdateHalfEdge for HalfEdge {
-    fn update_boundary(
-        &self,
-        update: impl FnOnce(CurveBoundary<Point<1>>) -> CurveBoundary<Point<1>>,
-    ) -> Self {
-        HalfEdge::new(
-            self.path(),
-            update(self.boundary()),
-            self.curve().clone(),
-            self.start_vertex().clone(),
-        )
-    }
-
     fn update_curve<T>(
         &self,
         update: impl FnOnce(&Handle<Curve>, &mut Core) -> T,
